test(query): Split parser tests into parser_test.go
